+++ src/abstraction/i-field.ts
@@ -1,0 +1,8 @@
+import { Cardinality } from '../enumeration';
+
+export interface IField {
+  name?: string;
+  alias?: string;
+  symbol?: string;
+  cardinality?: Cardinality;
+}