+++ src/abstraction/i-schema.ts
@@ -1,0 +1,8 @@
+import { Type } from './type';
+import { IField } from './i-field';
+
+export interface ISchema {
+  name?: string;
+  type?: Type;
+  fields?: IField[];
+}