+++ src/globalization/ schemas.ts
@@ -1,0 +1,3 @@
+import { ISchema } from '../abstraction';
+
+export const schemas = new Array<ISchema>();